move pkg/mcp/configz to pkg/mcp/configz/client
