+++ pkg/mcp/configz/client/assets.gen.go
