+++ pkg/mcp/configz/client/assets.go
