+++ pkg/mcp/configz/client/assets/templates/config.html
